+++ cli/lock.ml
@@ -104,7 +104,8 @@ let root_depexts local_opam_files =
     (fun _pkg (_version, opam_file) acc -> OpamFile.OPAM.depexts opam_file :: acc)
     local_opam_files []
 
-let run (`Repo repo) (`Recurse_opam recurse) (`Build_only build_only) (`Local_packages lp) () =
+let run (`Repo repo) (`Recurse_opam recurse) (`Build_only build_only) (`Local_packages lp)
+    (`Lockfile explicit_lockfile) () =
   let open Rresult.R.Infix in
   local_packages ~recurse ~explicit_list:lp repo >>= fun local_paths ->
   let local_packages =
@@ -113,7 +114,7 @@ let run (`Repo repo) (`Recurse_opam recurse) (`Build_only build_only) (`Local_pa
   in
   check_root_packages ~local_packages >>= fun () ->
   local_paths_to_opam_map local_paths >>= fun local_opam_files ->
-  Repo.lockfile ~local_packages repo >>= fun lockfile_path ->
+  Common.lockfile ~explicit_lockfile repo >>= fun lockfile_path ->
   calculate_opam ~build_only ~local_opam_files ~local_packages >>= fun package_summaries ->
   Common.Logs.app (fun l -> l "Calculating exact pins for each of them.");
   compute_duniverse ~package_summaries >>= resolve_ref >>= fun duniverse ->
@@ -185,6 +186,7 @@ let info =
 let term =
   let open Term in
   term_result
-    (const run $ Common.Arg.repo $ recurse_opam $ build_only $ packages $ Common.Arg.setup_logs ())
+    ( const run $ Common.Arg.repo $ recurse_opam $ build_only $ packages $ Common.Arg.lockfile
+    $ Common.Arg.setup_logs () )
 
 let cmd = (term, info)